+++ db/models/tag.go
@@ -5,7 +5,7 @@ import ()
 // Tag holds information about a tag in bag built by easy-store.
 // A cat in the hat from the hooka store.
 type Tag struct {
-	Id     int    `db:"id" form_options:"skip"`
+	Id     int64  `db:"id" form_options:"skip"`
 	BagId  *int   `db:"bag_id" form_widget:"hidden"`
 	Name   string `db:"name"`
 	Value  string `db:"value"`
@@ -13,7 +13,7 @@ type Tag struct {
 }
 
 // PrimaryKey() returns this object's Id, to conform to the Model interface.
-func (tag *Tag) PrimaryKey() int {
+func (tag *Tag) PrimaryKey() int64 {
 	return tag.Id
 }
